self-baker/requirements.txt
python-dotenv==1.0.1
openai==1.55.3
httpx==0.27.2
qdrant-client==1.9.1
requests==2.32.3
